+++ src/turnkeyml/common/build.py
@@ -137,12 +137,45 @@ def hash_model(model, model_type: ModelType, hash_params: bool = True):
         raise ValueError(msg)
 
 
-class Status(enum.Enum):
+class FunctionStatus(enum.Enum):
+    """
+    Status values that are assigned to stages, builds, benchmarks, and other
+    functionality to help the user understand whether that function completed
+    successfully or not.
+    """
+
+    # SUCCESSFUL means the stage/build/benchmark completed successfully.
+    SUCCESSFUL = "successful"
+
+    # ERROR means the stage/build/benchmark failed and threw some error that
+    # was caught by turnkey. You should proceed by looking at the build
+    # logs to see what happened.
+
+    ERROR = "error"
+
+    # KILLED means the build/benchmark failed because the system killed it. This can
+    # happen because of an out-of-memory (OOM), timeout, system shutdown, etc.
+    # You should proceed by re-running the build and keeping an eye on it to observe
+    # why it is being killed (e.g., watch the RAM utilization to diagnose an OOM).
+    KILLED = "killed"
+
+    # The NOT_STARTED status is applied to all stages/builds/benchmarks at startup.
+    # It will be replaced by one of the other status values if the stage/build/benchmark
+    # has a chance to start running.
+    # A value of NOT_STARTED in the report CSV indicates that the stage/build/benchmark
+    # never had a chance to start because turnkey exited before that functionality had
+    # a chance to start running.
     NOT_STARTED = "not_started"
-    PARTIAL_BUILD = "partial_build"
-    BUILD_RUNNING = "build_running"
-    SUCCESSFUL_BUILD = "successful_build"
-    FAILED_BUILD = "failed_build"
+
+    # INCOMPLETE indicates that a stage/build/benchmark started running and did not complete.
+    # Each stage, build, and benchmark are marked as INCOMPLETE when they start running.
+    # If you open the turnkey_stats.yaml file while the stage/build/benchmark
+    # is still running, the status will show as INCOMPLETE. If the stage/build/benchmark
+    # is killed without the chance to do any stats cleanup, the status will continue to
+    # show as INCOMPLETE in turnkey_stats.yaml.
+    # When the report CSV is created, any instance of an INCOMPLETE stage/build/benchmark
+    # status will be replaced by KILLED.
+    INCOMPLETE = "incomplete"
 
 
 # Create a unique ID from this run by hashing pid + process start time
@@ -242,7 +275,7 @@ class State:
     model_type: ModelType = ModelType.UNKNOWN
     uid: Optional[int] = None
     model_hash: Optional[int] = None
-    build_status: Status = Status.NOT_STARTED
+    build_status: FunctionStatus = FunctionStatus.NOT_STARTED
     expected_input_shapes: Optional[Dict[str, list]] = None
     expected_input_dtypes: Optional[Dict[str, list]] = None
     expected_output_names: Optional[List] = None
@@ -346,7 +379,7 @@ def load_state(
     try:
         # Special case for loading enums
         state_dict["model_type"] = ModelType(state_dict["model_type"])
-        state_dict["build_status"] = Status(state_dict["build_status"])
+        state_dict["build_status"] = FunctionStatus(state_dict["build_status"])
         state_dict["config"] = config_type(**state_dict["config"])
 
         state = state_type(**state_dict)